Split book details screen into components
